transform everything into a plugin
