+++ src/_includes/head.njk
@@ -53,11 +53,11 @@
     crossorigin="anonymous"
   />
 
-  <link rel="stylesheet" href="{{ '/scripts/main.entry.css' | asset | url}}" />
+  <link rel="stylesheet" href="{{ '/scripts/main.entry.css' | assetUrl }}" />
 
   <link rel="icon" href="/assets/images/favicon.png" />
 
-  {% set titleComplement %} {%if page.url != '/' %} |
+  {% set titleComplement %} {% if page.url != '/' %} |
   {{
     eleventyNavigation.key
   }}
