+++ src/_includes/layouts/lab.njk
@@ -6,11 +6,9 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
     <link rel="icon" href="/assets/images/favicon.png" />
-    <link
-      rel="stylesheet"
-      href="{{ '/styles/reset.entry.css' | asset | url }}"
-    />
-    <link rel="stylesheet" href="{{ './entry.css' | asset | url }}" />
+
+    {% assetStylesheet '/styles/reset.entry.css' %}
+    {% assetStylesheet './entry.css' %}
 
     {% set titleComplement = eleventyNavigation.title %}
 
@@ -19,6 +17,7 @@
 
   <body>
     {{ content | safe }}
-    <script async type="module" src="{{ './entry.js' | asset | url }}"></script>
+
+    {% assetScript './entry.js' %}
   </body>
 </html>
