+++ src/_includes/layouts/markdown.njk
@@ -1,11 +1,11 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
   <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <meta name="description" content="Christian Kaisermann - Resume">
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta name="description" content="Christian Kaisermann - Resume" />
 
-    <link rel="stylesheet" href="{{ '/styles/markdown.entry.css' | asset | url }}">
+    <link rel="stylesheet" href="{{ '/styles/markdown.entry.css' | assetUrl }}" />
 
     <title>kaisermann's resume</title>
   </head>
