+++ src/_includes/scripts/general.njk
@@ -16,6 +16,6 @@
 
 <script
   async
-  src="{{ '/scripts/main.entry.js' | asset | url }}"
+  src="{{ '/scripts/main.entry.js' | assetUrl }}"
   type="module"
 ></script>
