+++ src/assets.11ty.js
@@ -1,23 +1,1 @@
-class Asset {
-  data() {
-    return {
-      permalink: ({ asset }) => {
-        return asset.publicUrl;
-      },
-      pagination: {
-        addAllPagesToCollections: true,
-        alias: 'asset',
-        data: 'assets',
-        size: 1,
-      },
-      layout: '',
-      tags: ['_assets'],
-    };
-  }
-
-  render({ asset }) {
-    return asset.content;
-  }
-}
-
-module.exports = Asset;
+module.exports = require('./utils/assetManager').assetRenderer;
