+++ src/utils/assetManager/index.js
@@ -1,0 +1,7 @@
+const { assetBundler } = require('./bundler');
+const { AssetRenderer } = require('./renderer');
+const plugin = require('./plugin');
+
+exports.dataBundler = assetBundler;
+exports.assetRenderer = AssetRenderer;
+exports.plugin = plugin;
