+++ src/utils/assetManager/state.js
@@ -1,0 +1,11 @@
+let assets = [];
+
+exports.setAssets = (newAssets) => {
+  assets = newAssets;
+
+  return newAssets;
+};
+
+exports.getAssets = () => {
+  return assets;
+};
